form deciaml 일괄 변경 완료

+++ order/forms.py
@@ -32,8 +32,8 @@ class OrderForm(forms.Form):
                                  choices=list(Location.objects.values_list('code', 'codeName').order_by('code')),
                                  required=False)
     product = forms.ChoiceField(choices=list(ProductCode.objects.values_list('code', 'codeName')))  # TODO delete_State
-    amount = forms.FloatField(min_value=0)
-    amount_kg = forms.FloatField(min_value=0, widget=forms.HiddenInput())
+    amount = forms.DecimalField(decimal_places=2, max_digits=19, min_value=0)
+    amount_kg = forms.DecimalField(decimal_places=2, max_digits=19, min_value=0, widget=forms.HiddenInput())
     count = forms.IntegerField(min_value=0)
     price = forms.IntegerField(min_value=0)
     memo = forms.CharField(
