+++ product/forms.py
@@ -8,28 +8,28 @@ from .models import ProductCode, ProductEgg, ProductMaster
 
 class MainForm(forms.ModelForm):
     ymd = forms.CharField(max_length=8, widget=forms.HiddenInput())
-    total_loss_openEgg = forms.IntegerField(widget=forms.NumberInput(attrs=
+    total_loss_openEgg = forms.DecimalField(decimal_places=2, max_digits=19, widget=forms.NumberInput(attrs=
     {
         'placeholder': '할란',
         'data-toggle': 'tooltip',
         'data-placement': 'top',
         'title': '할란 로스량 입력'
     }))
-    total_loss_insert = forms.IntegerField(widget=forms.NumberInput(attrs=
+    total_loss_insert = forms.DecimalField(decimal_places=2, max_digits=19, widget=forms.NumberInput(attrs=
     {
         'placeholder': '투입',
         'data-toggle': 'tooltip',
         'data-placement': 'top',
         'title': '투입 로스량 입력'
     }))
-    total_loss_clean = forms.IntegerField(widget=forms.NumberInput(attrs=
+    total_loss_clean = forms.DecimalField(decimal_places=2, max_digits=19, widget=forms.NumberInput(attrs=
     {
         'placeholder': '살균',
         'data-toggle': 'tooltip',
         'data-placement': 'top',
         'title': '살균 로스량 입력'
     }))
-    total_loss_fill = forms.IntegerField(widget=forms.NumberInput(attrs=
+    total_loss_fill = forms.DecimalField(decimal_places=2, max_digits=19, widget=forms.NumberInput(attrs=
     {
         'placeholder': '충진',
         'data-toggle': 'tooltip',
@@ -43,54 +43,60 @@ class MainForm(forms.ModelForm):
 
 
 class StepOneForm(forms.Form):
-    할란_RawTank전란_01201 = forms.IntegerField(label='Raw Tank 전란', label_suffix='', help_text='01201',
+    할란_RawTank전란_01201 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 전란', label_suffix='',
+                                            help_text='01201',
                                             required=False, min_value=1)
-    할란사용_RawTank전란_01201 = forms.IntegerField(label='', required=False, min_value=1)
+    할란사용_RawTank전란_01201 = forms.DecimalField(decimal_places=2, max_digits=19, label='', required=False, min_value=1)
     메모_RawTank전란_01201_one = forms.CharField(
         label='memo',
         required=False,
         widget=forms.Textarea(attrs={'rows': 1}, )
     )
 
-    할란_RawTank난황_01202 = forms.IntegerField(label='Raw Tank 난황', label_suffix='', help_text='01202',
+    할란_RawTank난황_01202 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 난황', label_suffix='',
+                                            help_text='01202',
                                             required=False, min_value=1)
-    할란사용_RawTank난황_01202 = forms.IntegerField(label='', required=False, min_value=1)
+    할란사용_RawTank난황_01202 = forms.DecimalField(decimal_places=2, max_digits=19, label='', required=False, min_value=1)
     메모_RawTank난황_01202_one = forms.CharField(
         label='memo',
         required=False,
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    할란_RawTank난백_01203 = forms.IntegerField(label='Raw Tank 난백', label_suffix='', help_text='01203',
+    할란_RawTank난백_01203 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 난백', label_suffix='',
+                                            help_text='01203',
                                             required=False, min_value=1)
-    할란사용_RawTank난백_01203 = forms.IntegerField(label='', required=False, min_value=1)
+    할란사용_RawTank난백_01203 = forms.DecimalField(decimal_places=2, max_digits=19, label='', required=False, min_value=1)
     메모_RawTank난백_01203_one = forms.CharField(
         label='memo',
         required=False,
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    할란_RawTank등급란전란_01207 = forms.IntegerField(label='Raw Tank 등급란전란', label_suffix='', help_text='01207'
+    할란_RawTank등급란전란_01207 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 등급란전란', label_suffix='',
+                                               help_text='01207'
                                                , required=False, min_value=1)
-    할란사용_RawTank등급란전란_01207 = forms.IntegerField(label='', required=False, min_value=1)
+    할란사용_RawTank등급란전란_01207 = forms.DecimalField(decimal_places=2, max_digits=19, label='', required=False, min_value=1)
     메모_RawTank등급란전란_01207_one = forms.CharField(
         label='memo',
         required=False,
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    할란_RawTank등급란난황_01208 = forms.IntegerField(label='Raw Tank 등급란난황', label_suffix='', help_text='01208'
+    할란_RawTank등급란난황_01208 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 등급란난황', label_suffix='',
+                                               help_text='01208'
                                                , required=False, min_value=1)
-    할란사용_RawTank등급란난황_01208 = forms.IntegerField(label='', required=False, min_value=1)
+    할란사용_RawTank등급란난황_01208 = forms.DecimalField(decimal_places=2, max_digits=19, label='', required=False, min_value=1)
     메모_RawTank등급란난황_01208_one = forms.CharField(
         label='memo',
         required=False,
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    할란_RawTank등급란난백_01209 = forms.IntegerField(label='Raw Tank 등급란난백', label_suffix='', help_text='01209'
+    할란_RawTank등급란난백_01209 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 등급란난백', label_suffix='',
+                                               help_text='01209'
                                                , required=False, min_value=1)
-    할란사용_RawTank등급란난백_01209 = forms.IntegerField(label='', required=False, min_value=1)
+    할란사용_RawTank등급란난백_01209 = forms.DecimalField(decimal_places=2, max_digits=19, label='', required=False, min_value=1)
     메모_RawTank등급란난백_01209_one = forms.CharField(
         label='memo',
         required=False,
@@ -99,7 +105,7 @@ class StepOneForm(forms.Form):
 
 
 class StepTwoForm(forms.Form):
-    공정품투입_RawTank전란_01201 = forms.IntegerField(label='Raw Tank 전란', label_suffix='', help_text='01201'
+    공정품투입_RawTank전란_01201 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 전란', label_suffix='', help_text='01201'
                                                , required=False, min_value=1)
     메모_RawTank전란_01201_two = forms.CharField(
         label='',
@@ -107,7 +113,7 @@ class StepTwoForm(forms.Form):
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    공정품투입_RawTank난황_01202 = forms.IntegerField(label='Raw Tank 난황', label_suffix='', help_text='01202'
+    공정품투입_RawTank난황_01202 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 난황', label_suffix='', help_text='01202'
                                                , required=False, min_value=1)
     메모_RawTank난황_01202_two = forms.CharField(
         label='',
@@ -115,7 +121,7 @@ class StepTwoForm(forms.Form):
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    공정품투입_RawTank난백_01203 = forms.IntegerField(label='Raw Tank 난백', label_suffix='', help_text='01203'
+    공정품투입_RawTank난백_01203 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 난백', label_suffix='', help_text='01203'
                                                , required=False, min_value=1)
     메모_RawTank난백_01203_two = forms.CharField(
         label='',
@@ -123,7 +129,7 @@ class StepTwoForm(forms.Form):
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    공정품투입_PastTank전란_01204 = forms.IntegerField(label='Past Tank 전란', label_suffix='', help_text='01204'
+    공정품투입_PastTank전란_01204 = forms.DecimalField(decimal_places=2, max_digits=19, label='Past Tank 전란', label_suffix='', help_text='01204'
                                                 , required=False, min_value=1)
     메모_PastTank전란_01204_two = forms.CharField(
         label='',
@@ -131,7 +137,7 @@ class StepTwoForm(forms.Form):
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    공정품투입_PastTank난황_01205 = forms.IntegerField(label='Past Tank 난황', label_suffix='', help_text='01205'
+    공정품투입_PastTank난황_01205 = forms.DecimalField(decimal_places=2, max_digits=19, label='Past Tank 난황', label_suffix='', help_text='01205'
                                                 , required=False, min_value=1)
     메모_PastTank난황_01205_two = forms.CharField(
         label='',
@@ -139,7 +145,7 @@ class StepTwoForm(forms.Form):
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    공정품투입_PastTank난백_01206 = forms.IntegerField(label='Past Tank 난백', label_suffix='', help_text='01206'
+    공정품투입_PastTank난백_01206 = forms.DecimalField(decimal_places=2, max_digits=19, label='Past Tank 난백', label_suffix='', help_text='01206'
                                                 , required=False, min_value=1)
     메모_PastTank난백_01206_two = forms.CharField(
         label='',
@@ -147,7 +153,7 @@ class StepTwoForm(forms.Form):
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    공정품투입_RawTank등급란전란_01207 = forms.IntegerField(label='Raw Tank 등급란전란', label_suffix='', help_text='01207'
+    공정품투입_RawTank등급란전란_01207 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 등급란전란', label_suffix='', help_text='01207'
                                                   , required=False, min_value=1)
     메모_RawTank등급란전란_01207_two = forms.CharField(
         label='',
@@ -155,7 +161,7 @@ class StepTwoForm(forms.Form):
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    공정품투입_RawTank등급란난황_01208 = forms.IntegerField(label='Raw Tank 등급란난황', label_suffix='', help_text='01208'
+    공정품투입_RawTank등급란난황_01208 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 등급란난황', label_suffix='', help_text='01208'
                                                   , required=False, min_value=1)
     메모_RawTank등급란난황_01208_two = forms.CharField(
         label='',
@@ -163,7 +169,7 @@ class StepTwoForm(forms.Form):
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    공정품투입_RawTank등급란난백_01209 = forms.IntegerField(label='Raw Tank 등급란난백', label_suffix='', help_text='01209'
+    공정품투입_RawTank등급란난백_01209 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 등급란난백', label_suffix='', help_text='01209'
                                                   , required=False, min_value=1)
     메모_RawTank등급란난백_01209_two = forms.CharField(
         label='',
@@ -173,7 +179,7 @@ class StepTwoForm(forms.Form):
 
 
 class StepThreeForm(forms.Form):
-    공정품발생_RawTank전란_01201 = forms.IntegerField(label='Raw Tank 전란', label_suffix='', help_text='01201'
+    공정품발생_RawTank전란_01201 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 전란', label_suffix='', help_text='01201'
                                                , required=False, min_value=1)
     메모_RawTank전란_01201_three = forms.CharField(
         label='',
@@ -181,7 +187,7 @@ class StepThreeForm(forms.Form):
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    공정품발생_RawTank난황_01202 = forms.IntegerField(label='Raw Tank 난황', label_suffix='', help_text='01202'
+    공정품발생_RawTank난황_01202 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 난황', label_suffix='', help_text='01202'
                                                , required=False, min_value=1)
     메모_RawTank난황_01202_three = forms.CharField(
         label='',
@@ -189,7 +195,7 @@ class StepThreeForm(forms.Form):
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    공정품발생_RawTank난백_01203 = forms.IntegerField(label='Raw Tank 난백', label_suffix='', help_text='01203'
+    공정품발생_RawTank난백_01203 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 난백', label_suffix='', help_text='01203'
                                                , required=False, min_value=1)
     메모_RawTank난백_01203_three = forms.CharField(
         label='',
@@ -197,7 +203,7 @@ class StepThreeForm(forms.Form):
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    공정품발생_PastTank전란_01204 = forms.IntegerField(label='Past Tank 전란', label_suffix='', help_text='01204'
+    공정품발생_PastTank전란_01204 = forms.DecimalField(decimal_places=2, max_digits=19, label='Past Tank 전란', label_suffix='', help_text='01204'
                                                 , required=False, min_value=1)
     메모_PastTank전란_01204_three = forms.CharField(
         label='',
@@ -205,7 +211,7 @@ class StepThreeForm(forms.Form):
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    공정품발생_PastTank난황_01205 = forms.IntegerField(label='Past Tank 난황', label_suffix='', help_text='01205'
+    공정품발생_PastTank난황_01205 = forms.DecimalField(decimal_places=2, max_digits=19, label='Past Tank 난황', label_suffix='', help_text='01205'
                                                 , required=False, min_value=1)
     메모_PastTank난황_01205_three = forms.CharField(
         label='',
@@ -213,7 +219,7 @@ class StepThreeForm(forms.Form):
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    공정품발생_PastTank난백_01206 = forms.IntegerField(label='Past Tank 난백', label_suffix='', help_text='01206'
+    공정품발생_PastTank난백_01206 = forms.DecimalField(decimal_places=2, max_digits=19, label='Past Tank 난백', label_suffix='', help_text='01206'
                                                 , required=False, min_value=1)
     메모_PastTank난백_01206_three = forms.CharField(
         label='',
@@ -221,7 +227,7 @@ class StepThreeForm(forms.Form):
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    공정품발생_RawTank등급란전란_01207 = forms.IntegerField(label='Raw Tank 등급란전란', label_suffix='', help_text='01207'
+    공정품발생_RawTank등급란전란_01207 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 등급란전란', label_suffix='', help_text='01207'
                                                   , required=False, min_value=1)
     메모_RawTank등급란전란_01207_three = forms.CharField(
         label='',
@@ -229,7 +235,7 @@ class StepThreeForm(forms.Form):
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    공정품발생_RawTank등급란난황_01208 = forms.IntegerField(label='Raw Tank 등급란난황', label_suffix='', help_text='01208'
+    공정품발생_RawTank등급란난황_01208 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 등급란난황', label_suffix='', help_text='01208'
                                                   , required=False, min_value=1)
     메모_RawTank등급란난황_01208_three = forms.CharField(
         label='',
@@ -237,7 +243,7 @@ class StepThreeForm(forms.Form):
         widget=forms.Textarea(attrs={'rows': 1})
     )
 
-    공정품발생_RawTank등급란난백_01209 = forms.IntegerField(label='Raw Tank 등급란난백', label_suffix='', help_text='01209'
+    공정품발생_RawTank등급란난백_01209 = forms.DecimalField(decimal_places=2, max_digits=19, label='Raw Tank 등급란난백', label_suffix='', help_text='01209'
                                                   , required=False, min_value=1)
     메모_RawTank등급란난백_01209_three = forms.CharField(
         label='',
@@ -250,8 +256,8 @@ class StepFourForm(forms.Form):
     product = forms.ChoiceField(widget=Select2Widget,
                                 choices=list(ProductCode.objects.values_list('code', 'codeName')
                                              .filter(delete_state='N').order_by('code')), required=False)
-    amount = forms.FloatField(min_value=0)
-    amount_kg = forms.FloatField(min_value=0, widget=forms.HiddenInput())
+    amount = forms.DecimalField(decimal_places=2, max_digits=19, min_value=0)
+    amount_kg = forms.DecimalField(decimal_places=2, max_digits=19, min_value=0, widget=forms.HiddenInput())
     count = forms.IntegerField(min_value=0)
     memo = forms.CharField(
         label='',
@@ -265,12 +271,11 @@ StepFourFormSet = formset_factory(StepFourForm)
 class ProductOEMForm(forms.Form):
     product = forms.ChoiceField(widget=Select2Widget,
                                 choices=[('', '')] + list(ProductCode.objects.values_list('code', 'codeName')
-                                             .filter(delete_state='N').filter(oem='Y').order_by('code')))
+                                                          .filter(delete_state='N').filter(oem='Y').order_by('code')))
     location = forms.ChoiceField(widget=Select2Widget,
                                  choices=[('', '')] + list(Location.objects.values_list('code', 'codeName')
-                                              .filter(delete_state='N').filter(type='09').order_by('code')))
-    # amount = forms.FloatField(min_value=0)
-    # amount_kg = forms.FloatField(min_value=0, widget=forms.HiddenInput())
+                                     .filter(delete_state='N').filter(type='09').order_by(
+                                     'code')))
     count = forms.IntegerField(min_value=0)
     purchaseSupplyPrice = forms.IntegerField(min_value=0)
     purchaseVat = forms.IntegerField(min_value=0, required=False)
